src/wallet/requirements.txt
base58==2.1.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
ecdsa==0.18.0
hdwallet==2.1.1
idna==3.3
mnemonic==0.20
pycryptodome==3.15.0
pysha3==1.0.2
requests==2.28.1
six==1.16.0
urllib3==1.26.12
